fix(advisers): planar clearance rejection, CMC material sort, scoring weight and direction fixes
- CoilAdviser: B19 planar clearance violation only continued the section
  loop and never rejected the wire combination; integer division before
  ceil in per-pattern result budget.
- CoreAdviserMaterials: impedance-driven material selection sorted
  ascending, picking the LOWEST permeability ferrites; fan-out loops could
  index past the evaluations vector.
- WireAdviser: integer division when prorating outer area per section.
- MagneticFilterArea/EnergyCost: toroidal margin scaled usable window area
  by marginAngle/360 instead of (360-marginAngle)/360; allowNotFit overflow
  scoring rewarded bigger overflows (now 1-extra); MagneticFilterCost
  accumulated double costs into int and dereferenced an unchecked optional.
- Utils normalize_scoring: all-equal branch added +1 ignoring the filter
  weight (corrupting cross-filter weighting); added empty-input guards.

+++ src/advisers/WireAdviser.cpp
@@ -646,7 +646,7 @@ void WireAdviser::set_maximum_area_proportion(std::vector<std::pair<Winding, dou
         }
         auto neededOuterAreaNoCompact = wire.get_maximum_outer_width() * wire.get_maximum_outer_height();
 
-        neededOuterAreaNoCompact *= (*unfilteredCoils)[coilIndex].first.get_number_parallels() * (*unfilteredCoils)[coilIndex].first.get_number_turns() / numberSections;
+        neededOuterAreaNoCompact *= (*unfilteredCoils)[coilIndex].first.get_number_parallels() * (*unfilteredCoils)[coilIndex].first.get_number_turns() / static_cast<double>(numberSections);
 
         double areaProportion = neededOuterAreaNoCompact / sectionArea;
         _maximumOuterAreaProportion = std::max(_maximumOuterAreaProportion, areaProportion);
